requirements.txt
langchain
langgraph
python-dotenv
langchain-community
google-generativeai
pymysql
ipykernel
streamlit

-e .